+++ Plutarch/Rational.hs
@@ -1,5 +1,5 @@
 {-# LANGUAGE FlexibleInstances #-}
-{-# OPTIONS_GHC -Wno-orphans #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module Plutarch.Rational (
   PRational (PRational),
@@ -13,7 +13,8 @@ module Plutarch.Rational (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Builtin.Bool (pcond, pif)
+import Generics.SOP qualified as SOP
+import Plutarch.Builtin.Bool (PBool, pcond, pif)
 import Plutarch.Builtin.Data (PAsData, PBuiltinList, PData, ppairDataBuiltin)
 import Plutarch.Builtin.Integer (PInteger)
 import Plutarch.Internal.Eq (PEq ((#==)))
@@ -37,29 +38,32 @@ import Plutarch.Internal.Lift (
  )
 import Plutarch.Internal.ListLike (phead, pnil, ptail)
 import Plutarch.Internal.Numeric (
-  PNum (pabs, pfromInteger, pnegate, psignum, (#*), (#+), (#-)),
+  PAdditiveGroup (pnegate, pscaleInteger, (#-)),
+  PAdditiveMonoid (pzero),
+  PAdditiveSemigroup (pscalePositive, (#+)),
+  PIntegralDomain (pabs, psignum),
+  PMultiplicativeMonoid (pone),
+  PMultiplicativeSemigroup (ppowPositive, (#*)),
+  PPositive,
+  PRing (pfromInteger),
   pdiv,
   pmod,
   pquot,
+  ptryPositive,
  )
 import Plutarch.Internal.Ord (
-  POrd (pmax, pmin, (#<), (#<=)),
+  POrd ((#<), (#<=)),
  )
 import Plutarch.Internal.Other (pto)
 import Plutarch.Internal.PLam (plam)
-import Plutarch.Internal.PlutusType (
-  DerivePlutusType (DPTStrat),
-  PlutusType,
-  pcon,
-  pmatch,
- )
-import Plutarch.Internal.ScottEncoding (PlutusTypeScott)
+import Plutarch.Internal.PlutusType (PlutusType, pcon, pmatch)
 import Plutarch.Internal.Show (PShow, pshow, pshow')
 import Plutarch.Internal.Term (
   S,
   Term,
   phoistAcyclic,
   plet,
+  punsafeBuiltin,
   (#),
   (#$),
   (:-->),
@@ -71,9 +75,10 @@ import Plutarch.Internal.TermCont (
  )
 import Plutarch.Internal.TryFrom (PTryFrom (PTryFromExcess, ptryFrom'), ptryFrom)
 import Plutarch.Pair (PPair (PPair))
-import Plutarch.Positive (PPositive, ptryPositive)
+import Plutarch.Repr.SOP (DeriveAsSOPRec (DeriveAsSOPRec))
 import Plutarch.Trace (ptraceInfoError)
 import Plutarch.Unsafe (punsafeCoerce, punsafeDowncast)
+import PlutusCore qualified as PLC
 import PlutusTx.Ratio qualified as PlutusTx
 
 {- | A Scott-encoded rational number, with a guaranteed positive denominator
@@ -90,10 +95,9 @@ numbers that isn't just passing them around, you want to use (or convert to)
 data PRational s
   = PRational (Term s PInteger) (Term s PPositive)
   deriving stock (Generic)
-  deriving anyclass (PlutusType)
+  deriving anyclass (SOP.Generic)
 
-instance DerivePlutusType PRational where
-  type DPTStrat _ = PlutusTypeScott
+deriving via (DeriveAsSOPRec PRational) instance PlutusType PRational
 
 -- | @since WIP
 instance PLiftable PRational where
@@ -107,7 +111,7 @@ instance PLiftable PRational where
         d = PlutusTx.denominator r
      in mkPLifted . pcon $
           if
-            | n == 0 -> PRational 0 1
+            | n == 0 -> PRational pzero pone
             | d < 0 -> PRational (pconstant . negate $ n) . punsafeCoerce . pconstant @PInteger . negate $ d
             | otherwise -> PRational (pconstant n) . punsafeCoerce . pconstant @PInteger $ d
   {-# INLINEABLE fromPlutarchRepr #-}
@@ -119,15 +123,125 @@ instance PLiftable PRational where
     pure . PlutusTx.unsafeRatio n $ d
 
 instance PEq PRational where
-  l' #== r' =
+  {-# INLINEABLE (#==) #-}
+  l' #== r' = inner # l' # r'
+    where
+      inner :: forall (s :: S). Term s (PRational :--> PRational :--> PBool)
+      inner = phoistAcyclic $ plam $ \l r ->
+        cmpHelper # punsafeBuiltin PLC.EqualsInteger # l # r
+
+instance POrd PRational where
+  {-# INLINEABLE (#<=) #-}
+  l' #<= r' = inner # l' # r'
+    where
+      inner :: forall (s :: S). Term s (PRational :--> PRational :--> PBool)
+      inner = phoistAcyclic $ plam $ \l r ->
+        cmpHelper # punsafeBuiltin PLC.LessThanEqualsInteger # l # r
+  {-# INLINEABLE (#<) #-}
+  l' #< r' = inner # l' # r'
+    where
+      inner :: forall (s :: S). Term s (PRational :--> PRational :--> PBool)
+      inner = phoistAcyclic $ plam $ \l r ->
+        cmpHelper # punsafeBuiltin PLC.LessThanInteger # l # r
+
+-- | @since WIP
+instance PAdditiveSemigroup PRational where
+  {-# INLINEABLE (#+) #-}
+  x' #+ y' =
     phoistAcyclic
-      ( plam $ \l r ->
-          pmatch l $ \(PRational ln ld) ->
-            pmatch r $ \(PRational rn rd) ->
-              pto rd * ln #== rn * pto ld
+      ( plam $ \x y -> unTermCont $ do
+          PRational xn xd' <- tcont $ pmatch x
+          PRational yn yd' <- tcont $ pmatch y
+          xd <- tcont $ plet xd'
+          yd <- tcont $ plet yd'
+          pure $ preduce' # (xn * pto yd + yn * pto xd) # pto (xd #* yd)
       )
-      # l'
-      # r'
+      # x'
+      # y'
+  {-# INLINEABLE pscalePositive #-}
+  pscalePositive = phoistAcyclic $ plam $ \x p ->
+    pmatch x $ \(PRational xn xd) ->
+      preduce' # (xn #* pto p) # pto xd
+
+-- | @since WIP
+instance PAdditiveMonoid PRational where
+  {-# INLINEABLE pzero #-}
+  pzero = pcon . PRational pzero $ pone
+
+-- | @since WIP
+instance PAdditiveGroup PRational where
+  {-# INLINEABLE pnegate #-}
+  pnegate =
+    phoistAcyclic $
+      plam $ \x ->
+        pmatch x $ \(PRational xn xd) ->
+          pcon $ PRational (pnegate # xn) xd
+  {-# INLINEABLE (#-) #-}
+  x' #- y' =
+    phoistAcyclic
+      ( plam $ \x y -> unTermCont $ do
+          PRational xn xd' <- tcont $ pmatch x
+          PRational yn yd' <- tcont $ pmatch y
+          xd <- tcont $ plet xd'
+          yd <- tcont $ plet yd'
+          pure $ preduce' # (xn * pto yd - yn * pto xd) # pto (xd #* yd)
+      )
+      # x'
+      # y'
+  {-# INLINEABLE pscaleInteger #-}
+  pscaleInteger = phoistAcyclic $ plam $ \x e ->
+    pmatch x $ \(PRational xn xd) ->
+      preduce' # (xn #* e) # pto xd
+
+{-
+-- | @since WIP
+instance PAbs PRational where
+  {-# INLINEABLE pabs #-}
+-}
+
+-- | @since WIP
+instance PMultiplicativeSemigroup PRational where
+  {-# INLINEABLE (#*) #-}
+  x' #* y' =
+    phoistAcyclic
+      ( plam $ \x y -> unTermCont $ do
+          PRational xn xd <- tcont $ pmatch x
+          PRational yn yd <- tcont $ pmatch y
+          pure $ preduce' # (xn * yn) # pto (xd #* yd)
+      )
+      # x'
+      # y'
+  {-# INLINEABLE ppowPositive #-}
+  ppowPositive = phoistAcyclic $ plam $ \x p ->
+    pmatch x $ \(PRational xn xd) ->
+      pcon . PRational (ppowPositive # xn # p) $ ppowPositive # xd # p
+
+-- | @since WIP
+instance PMultiplicativeMonoid PRational where
+  {-# INLINEABLE pone #-}
+  pone = pcon . PRational pone $ pone
+
+-- | @since WIP
+instance PRing PRational where
+  {-# INLINEABLE pfromInteger #-}
+  pfromInteger n = pcon $ PRational (fromInteger n) pone
+
+-- | @since WIP
+instance PIntegralDomain PRational where
+  {-# INLINEABLE pabs #-}
+  pabs =
+    phoistAcyclic $
+      plam $ \x ->
+        pmatch x $ \(PRational xn xd) ->
+          pcon $ PRational (pabs # xn) xd
+  {-# INLINEABLE psignum #-}
+  psignum = phoistAcyclic $ plam $ \x ->
+    pmatch x $ \(PRational n _) ->
+      pcond
+        [ (n #== 0, pzero)
+        , (n #<= 0, pcon . PRational (pconstant (-1)) $ pone)
+        ]
+        pone
 
 -- | @since WIP
 instance Fractional (Term s PRational) where
@@ -141,13 +255,7 @@ instance Fractional (Term s PRational) where
             pif
               (denm #== 0)
               (ptraceInfoError "Cannot divide by zero")
-              ( plet (xn * pto yd) $ \numm ->
-                  preduce
-                    #$ pif
-                      (denm #< 0)
-                      (pcon $ PRational (pnegate #$ numm) (punsafeCoerce $ pnegate # denm))
-                      (pcon $ PRational numm (punsafeCoerce denm))
-              )
+              (preduce' # (xn * pto yd) # denm)
   {-# INLINEABLE recip #-}
   recip x = inner # x
     where
@@ -155,7 +263,7 @@ instance Fractional (Term s PRational) where
       inner = phoistAcyclic $ plam $ \x -> pmatch x $ \(PRational xn xd) ->
         pcond
           [ (xn #== 0, ptraceInfoError "attempted to construct the reciprocal of zero")
-          , (xn #< 0, pcon $ PRational (pnegate #$ pto xd) (punsafeCoerce $ pnegate # xn))
+          , (xn #<= 0, pcon $ PRational (pnegate #$ pto xd) (punsafeCoerce $ pnegate # xn))
           ]
           (pcon $ PRational (pto xd) (punsafeCoerce xn))
   {-# INLINEABLE fromRational #-}
@@ -182,116 +290,20 @@ instance PTryFrom PData (PAsData PRational) where
     res <- tcont . plet $ ptryPositive # denm
     pure (punsafeCoerce opq, res)
 
-instance POrd PRational where
-  {-# INLINEABLE (#<=) #-}
-  l' #<= r' =
-    phoistAcyclic
-      ( plam $ \l r -> unTermCont $ do
-          PRational ln ld <- tcont $ pmatch l
-          PRational rn rd <- tcont $ pmatch r
-          pure $ pto rd * ln #<= rn * pto ld
-      )
-      # l'
-      # r'
-  {-# INLINEABLE (#<) #-}
-  l' #< r' =
-    phoistAcyclic
-      ( plam $ \l r -> unTermCont $ do
-          PRational ln ld <- tcont $ pmatch l
-          PRational rn rd <- tcont $ pmatch r
-          pure $ pto rd * ln #< rn * pto ld
-      )
-      # l'
-      # r'
-
-instance PNum PRational where
-  {-# INLINEABLE (#+) #-}
-  x' #+ y' =
-    phoistAcyclic
-      ( plam $ \x y -> unTermCont $ do
-          PRational xn xd' <- tcont $ pmatch x
-          PRational yn yd' <- tcont $ pmatch y
-          xd <- tcont $ plet xd'
-          yd <- tcont $ plet yd'
-          pure
-            $ preduce
-              #$ pcon
-            $ PRational (xn * pto yd + yn * pto xd)
-            $ punsafeDowncast
-            $ pto xd * pto yd
-      )
-      # x'
-      # y'
-  {-# INLINEABLE (#-) #-}
-  x' #- y' =
-    phoistAcyclic
-      ( plam $ \x y -> unTermCont $ do
-          PRational xn xd' <- tcont $ pmatch x
-          PRational yn yd' <- tcont $ pmatch y
-          xd <- tcont $ plet xd'
-          yd <- tcont $ plet yd'
-          pure
-            $ preduce
-              #$ pcon
-            $ PRational (xn * pto yd - yn * pto xd)
-            $ punsafeDowncast
-            $ pto xd * pto yd
-      )
-      # x'
-      # y'
-  {-# INLINEABLE (#*) #-}
-  x' #* y' =
-    phoistAcyclic
-      ( plam $ \x y -> unTermCont $ do
-          PRational xn xd <- tcont $ pmatch x
-          PRational yn yd <- tcont $ pmatch y
-          pure
-            $ preduce
-              #$ pcon
-            $ PRational (xn * yn)
-            $ punsafeDowncast
-            $ pto xd * pto yd
-      )
-      # x'
-      # y'
-  {-# INLINEABLE pnegate #-}
-  pnegate =
-    phoistAcyclic $
-      plam $ \x ->
-        pmatch x $ \(PRational xn xd) ->
-          pcon $ PRational (negate xn) xd
-  {-# INLINEABLE pabs #-}
-  pabs =
-    phoistAcyclic $
-      plam $ \x ->
-        pmatch x $ \(PRational xn xd) ->
-          pcon $ PRational (abs xn) xd
-  {-# INLINEABLE psignum #-}
-  psignum = phoistAcyclic $ plam $ \x ->
-    pmatch x $ \(PRational n _) ->
-      pcond
-        [ (n #== 0, 0)
-        , (n #< 0, -1)
-        ]
-        1
-  {-# INLINEABLE pfromInteger #-}
-  pfromInteger n = pcon $ PRational (fromInteger n) 1
-
 preduce :: Term s (PRational :--> PRational)
-preduce = phoistAcyclic $
-  plam $ \x -> unTermCont $ do
-    PRational xn xd' <- tcont $ pmatch x
-    xd <- tcont . plet $ pto xd'
-    r <- tcont . plet $ pgcd # xn # xd
-    s <- tcont . plet $ psignum # xd
-    pure . pcon $ PRational (s * pdiv # xn # r) $ punsafeDowncast $ s * pdiv # xd # r
+preduce = phoistAcyclic $ plam $ \x ->
+  pmatch x $ \(PRational n d) -> preduce' # n # pto d
 
 pgcd :: Term s (PInteger :--> PInteger :--> PInteger)
 pgcd = phoistAcyclic $
   plam $ \x' y' -> unTermCont $ do
     x <- tcont . plet $ pabs # x'
     y <- tcont . plet $ pabs # y'
-    pure $ pgcd' # pmax x y #$ pmin x y
+    pure $
+      pif
+        (x #<= y)
+        (pgcd' # y # x)
+        (pgcd' # x # y)
 
 -- assumes inputs are non negative and a >= b
 pgcd' :: Term s (PInteger :--> PInteger :--> PInteger)
@@ -310,7 +322,7 @@ pdenominator :: Term s (PRational :--> PPositive)
 pdenominator = phoistAcyclic $ plam $ \x -> pmatch x $ \(PRational _ d) -> d
 
 pfromInteger :: Term s (PInteger :--> PRational)
-pfromInteger = phoistAcyclic $ plam $ \n -> pcon $ PRational n 1
+pfromInteger = phoistAcyclic $ plam $ \n -> pcon $ PRational n pone
 
 pround :: Term s (PRational :--> PInteger)
 pround = phoistAcyclic $
@@ -340,3 +352,23 @@ pproperFraction = phoistAcyclic $
   plam $ \x ->
     plet (ptruncate # x) $ \q ->
       pcon $ PPair q (x - Plutarch.Rational.pfromInteger # q)
+
+-- Helpers
+
+cmpHelper ::
+  forall (s :: S).
+  Term s ((PInteger :--> PInteger :--> PBool) :--> PRational :--> PRational :--> PBool)
+cmpHelper = phoistAcyclic $ plam $ \f l r ->
+  pmatch l $ \(PRational ln ld) ->
+    pmatch r $ \(PRational rn rd) ->
+      f # (pto rd * ln) # (rn * pto ld)
+
+-- Assumes d is not zero
+preduce' :: forall (s :: S). Term s (PInteger :--> PInteger :--> PRational)
+preduce' = phoistAcyclic $ plam $ \n d' ->
+  plet d' $ \d ->
+    plet (pgcd # n # d) $ \r ->
+      pif
+        (d #<= 0)
+        (pcon $ PRational (pnegate # (pdiv # n # r)) $ punsafeDowncast (pnegate # (pdiv # d # r)))
+        (pcon $ PRational (pdiv # n # r) $ punsafeDowncast (pdiv # d # r))